Feat: FarmUpgradePanel UI 기능 구현, FarmUpgradePanel 레이아웃 조정, DTO 추가

+++ Assets/2.Scripts/FarmData.cs
@@ -25,7 +25,7 @@ public class FarmData
     [Header("최대 농장 레벨")]
     public int maxFarmLevel;
     [Header("농장 기본 골드")]
-    public int basicFarmGold;
+    public long basicFarmGold;
     [Header("레벨별 농장 골드 증가율")]
     public float farmGoldPer;
 
@@ -37,7 +37,7 @@ public class FarmData
     [Header("레벨당 생산 확률 증가(%)")]
     public float autoProduceChancePer;
     [Header("자동 생산 확률 기본 골드")]
-    public int basicAutoProduceChanceGold;
+    public long basicAutoProduceChanceGold;
     [Header("레벨별 자동 생산 확률 골드 증가율")]
     public float autoProduceChanceGoldPer;
 
@@ -50,7 +50,7 @@ public class FarmData
     [Header("레벨당 생산 시간 감소(sec)")]
     public float produceCooldownPer;
     [Header("생산 시간 감소 기본 골드")]
-    public int basicProduceCooldownGold;
+    public long basicProduceCooldownGold;
     [Header("레벨별 생산 시간 감소 골드 증가율")]
     public float produceCooldownGoldPer;
